requirements.txt
streamlit
supabase
google-generativeai
requests